fix(agent): remove redundant u64 conversion in trace token counting
The `usage.input_tokens` and `usage.cache_read_tokens` fields were already u64, making the `u64::from` wrapping unnecessary. This change removes the redundant conversions to keep the code cleaner and avoid potential confusion about the types involved.

Auto-committed-on: dragonfly

+++ src/agent/trace.rs
@@ -193,10 +193,10 @@ impl EventListener for RunTracer {
                 if let Some(usage) = usage {
                     self.state
                         .input_tokens
-                        .fetch_add(u64::from(usage.input_tokens), Ordering::Relaxed);
+                        .fetch_add(usage.input_tokens, Ordering::Relaxed);
                     self.state
                         .cached_tokens
-                        .fetch_add(u64::from(usage.cache_read_tokens), Ordering::Relaxed);
+                        .fetch_add(usage.cache_read_tokens, Ordering::Relaxed);
                 }
                 let detail = usage.as_ref().map_or_else(String::new, |usage| {
                     format!(
